Trying to better handle closing flow of the Iterator.
Next() should be the only thing that drives error handling on the consuming
side.  It should empty all the items in the case of a proper termination
(there's a bug regarding this, left in this first commit).

No one should interfere with the internal `context.Context` of the iterator,
so removed it from being exposed.

No need to handle `Close()` either, as `Next()` should do all the clean-up.
Assumption here is that you should definitely loop through `Next()`, always.
OK what if you want to break early for some reasons? Is that where we should
have a `.Close()` ?