Preemptively push batch that would exceed allocated max batch size
This is now how Bigkv and TiKV behaves, i.e. that they would flush a current batch in
progress if the new element would bust the size or op limit.

This is especially important for TiKV which default batch max size is "relatively small"
at 8MB. Even when set to a batch size of 7MB, when an element in itself is > 1MB, it will
actually cause an error later on in the TiKV client.

By checking first if the next item would need a flush, we can flush the current batch and
add the item to after it, reducing max write size error to the single case where a single
element is >8MB.